+++ tests/test_kafka_consumer.py
@@ -9,7 +9,7 @@ def test_parse_dataset_message():
     message = {
         'service': 'udata',
         'meta': {
-            'message_type': 'index',
+            'message_type': 'dataset.index',
             'index': 'dataset'
         },
         'data': {
@@ -46,7 +46,7 @@ def test_parse_dataset_message():
         }
     }
     val_utf8 = json.dumps(message)
-    message_type, index_name, data = parse_message('dataset', val_utf8)
+    message_type, index_name, data = parse_message(val_utf8)
 
     assert message_type == 'index'
     assert index_name == 'dataset'
@@ -83,7 +83,7 @@ def test_parse_reuse_message():
     message = {
         'service': 'udata',
         'meta': {
-            'message_type': 'index',
+            'message_type': 'reuse.index',
             'index': 'reuse'
         },        'data': {
             "id": "5cc2dfbe8b4c414c91ffc46d",
@@ -108,7 +108,7 @@ def test_parse_reuse_message():
         }
     }
     val_utf8 = json.dumps(message)
-    message_type, index_name, data = parse_message('reuse', val_utf8)
+    message_type, index_name, data = parse_message(val_utf8)
 
     assert message_type == 'index'
     assert index_name == 'reuse'
@@ -136,7 +136,7 @@ def test_parse_organization_message():
     message = {
         'service': 'udata',
         'meta': {
-            'message_type': 'index',
+            'message_type': 'organization.index',
             'index': 'organization'
         },        'data': {
             "id": "534fff75a3a7292c64a77de4",
@@ -154,7 +154,7 @@ def test_parse_organization_message():
         }
     }
     val_utf8 = json.dumps(message)
-    message_type, index_name, data = parse_message('organization', val_utf8)
+    message_type, index_name, data = parse_message(val_utf8)
 
     assert message_type == 'index'
     assert index_name == 'organization'